Prep for build with flit
Only cosmetic updates here with the exception of the `pyproject.toml`
file, which supplants the old SetupTools `setup.py` and `setup.cfg`.

Apparently my old build would not import via `pip` in Python 3.11
while testing for Curator. So here we go with an updated version,
cleaned up, and ready for building with flit.