Move the try/except block for Docker logging
It needs to be out one more level in order to catch the file permissions check.
It turned out to be a faulty expectation that it would not test for write permissions until it actually attempted to write, instead of immediately upon calling, `with open(fpath, 'wb+', ...)`

Additional change is cosmetic Pylint stuff.

All tests pass.